:recycle: [ Refactor ] QuestionListPage 상수화 리팩토링 작업
refactor
