+++ src/components/organisms/PageNationNumbers/PageNationNumbers.jsx
@@ -1,18 +1,18 @@
 import styled from 'styled-components';
 import PageNationButton from '../../atoms/Button/PageNationButton/PageNationButton';
+import { PAGE_NATION_NUMBER_COUNT } from '../../../constants/page_nation';
 
 const PageNationNumbers = ({ start, currentPage, onClick }) =>
-  Array(5)
+  Array(PAGE_NATION_NUMBER_COUNT)
     .fill()
     .map((_, i) => {
       const page = start + i;
-      const isSelected = page === currentPage;
       return (
         <Container>
           <PageNationButton
             key={page}
             page={page}
-            isSelected={isSelected}
+            isSelected={page === currentPage}
             onClick={onClick}
           />
         </Container>
