--- src/constants/constants.js
@@ -1,9 +1,0 @@
-export const NEXT_ARROW_BACKGROUND = `url("data:image/svg+xml,%3csvg width='100%25' height='100%25' xmlns='http://www.w3.org/2000/svg'%3e%3crect width='100%25' height='100%25' fill='none' rx='17' ry='17' stroke='%23818181FF' stroke-width='1' stroke-dasharray='9' stroke-dashoffset='29' stroke-linecap='square'/%3e%3c/svg%3e")`;
-
-export const SORT_TIME = 'time';
-
-export const SORT_TIME_KOREAN = '시간순';
-
-export const SORT_NAME = 'name';
-
-export const SORT_NAME_KOREAN = '이름순';
