+++ src/constants/question_list_page_sort.js
@@ -1,0 +1,7 @@
+export const SORT_TIME = 'time';
+
+export const SORT_TIME_KOREAN = '시간순';
+
+export const SORT_NAME = 'name';
+
+export const SORT_NAME_KOREAN = '이름순';
