+++ src/constants/svg_Img_Url.js
@@ -1,0 +1,3 @@
+const NEXT_ARROW_BACKGROUND = `url("data:image/svg+xml,%3csvg width='100%25' height='100%25' xmlns='http://www.w3.org/2000/svg'%3e%3crect width='100%25' height='100%25' fill='none' rx='17' ry='17' stroke='%23818181FF' stroke-width='1' stroke-dasharray='9' stroke-dashoffset='29' stroke-linecap='square'/%3e%3c/svg%3e")`;
+
+export default NEXT_ARROW_BACKGROUND;
