+++ src/hooks/useQuestionListPageData.jsx
@@ -1,13 +1,13 @@
 import { useEffect, useState } from 'react';
 import requestApi from '../utils/requestApi';
-import { SORT_TIME } from '../constants/constants';
+import { SORT_TIME } from '../constants/question_list_page_sort';
 
 /**
  *
  * @returns data, isLoading, error, totalPage, currentPage, handleLeftArrowClick, handleRightArrowClick, handleSortButtonClick, handlePageClick
  */
 
-const useGetQuestionPageData = () => {
+const useQuestionListPageData = () => {
   const [currentPage, setCurrentPage] = useState(1);
   const limit = 8;
   const offset = (currentPage - 1) * limit;
@@ -79,4 +79,4 @@ const useGetQuestionPageData = () => {
   };
 };
 
-export default useGetQuestionPageData;
+export default useQuestionListPageData;
